Connect existing validation middlewares to listings and reviews routes
