+++ utils/ExpressError.js
@@ -1,0 +1,9 @@
+class ExpressError extends Error {
+  constructor(statusCode, message) {
+    super();
+    this.statusCode = statusCode;
+    this.message = message;
+  }
+}
+
+module.exports = ExpressError;
